refactor: reorganize OAuth-related imports and rename hooks for clarity
