+++ src/app/(private)/profile/_hooks/useLogout.ts
